step 1: move sql models into ell.stores
this does not cut core's dependency on studio/sqlmodel yet but moves in that direction
