+++ src/ell/types/__init__.py
@@ -3,5 +3,4 @@ The primary types used in ell
 """
 
 from ell.types.message import *
-from ell.types.studio import *
 from ell.types._lstr import *
